refactor(molecules): merge SfSelect
