--- packages/vue/src/components/molecules/SfSelect/SfSelect.html
@@ -1,55 +1,0 @@
-<div
-  @click="toggle($event)"
-  @blur="closeHandler"
-  @keyup.space="openHandler"
-  @keyup.up="move(-1)"
-  @keyup.down="move(1)"
-  @keyup.enter="enter($event)"
-  :aria-expanded="open ? 'true' : 'false'"
-  :aria-owns="'lbox_'+_uid"
-  aria-autocomplete="none"
-  role="combobox"
-  tabindex="0"
-  :class="{
-  'sf-select--is-active': isActive,
-  'sf-select--is-selected': isSelected,
-  'sf-select--is-required': required
-  }"
-  class="sf-select">
-  <div style="position: relative">
-    <div
-      v-html="html"
-      class="sf-select__selected sf-select-option"></div>
-    <div
-      v-if="label"
-      class="sf-select__label">
-      <slot name="label">{{label}}</slot>
-    </div>
-    <SfOverlay :visible="open" @click="" class="sf-select__overlay"/>
-    <transition name="sf-select">
-      <div v-show="open" class="sf-select__dropdown">
-        <!--  sf-select__option -->
-        <ul
-          :style="{ maxHeight }"
-          class="sf-select__options">
-          <slot/>
-        </ul>
-        <SfButton
-          @click="closeHandler"
-          ref="cancel"
-          class="sf-select__cancel sf-button--full-width">
-          Cancel
-        </SfButton>
-      </div>
-    </transition>
-  </div>
-  <div
-    v-if="(valid !== undefined)"
-    class="sf-select__error-message">
-    <transition name="fade">
-      <span v-if="(valid === false)">
-        <slot name="errorMessage">{{errorMessage}}</slot>
-      </span>
-    </transition>
-  </div>
-</div>
