--- packages/vue/src/components/molecules/SfSelect/_internal/SfSelectOption.html
@@ -1,9 +1,0 @@
-<li
-  @click="clicked"
-  class="sf-select-option"
-  :class="{'sf-select-option--active': selected === value}"
-  role="option"
-  :aria-selected="selected === value ? 'true' : 'false'">
-  <!-- @slot -->
-  <slot />
-</li>
